Factor out command output creation
